fix(ci): split projector files and drop deleted-task state
Keep statussummary under the file-length lint limit. On verified
missing-task queue-status events drop projectionState, suppress only
not-found/FK failures, and leave transient errors to retry.
