+++ apps/backend/internal/task/statussummary/constants.go
@@ -8,6 +8,14 @@ const (
 
 	messageTypeClarificationRequest = "clarification_request"
 	messageTypePermissionRequest    = "permission_request"
+	messageTypeError                = "error"
+	messageTypeStatus               = "status"
+	messageTypeAgent                = "agent"
+	messageTypeStatusUpdate         = "status_update"
+
+	sessionStateStarting        = "STARTING"
+	sessionStateRunning         = "RUNNING"
+	sessionStateWaitingForInput = "WAITING_FOR_INPUT"
 
 	activityBackground = "background"
 	activityGenerating = "generating"
